tests/wpt/tests/html/editing/dnd/images/022.xhtml

<?xml version="1.0" encoding="utf-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/test-helper.js"></script>
<head>
  <title>Dropping image on vertical iframe scrollbar</title>
  <style type="text/css">
    p:first-child {
      padding-left: 1em;
    }

    iframe {
      height: 6em;
      width: 3em;
    }

    img {
      width: 5px;
      height: 5px;
    }
  </style>
  <script>
    function dragImage() {
      event.dataTransfer.effectAllowed = 'copy';
    }
  </script>
</head>
<body ondragenter="event.preventDefault()" ondragover="return false">
  <p><img
      src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAEElEQVR42mNgaGD4D8YwBgAw9AX9Y9zBwwAAAABJRU5ErkJggg=="
      alt="PNG green pixel" ondragstart="dragImage()" /></p>
  <p>Drag little square above and drop it on vertical scrollbar. Word PASS should appear once you drop it.</p>
  <iframe src="helper-drop-vertical-scrollbar.xhtml">XHTML document</iframe>
  <script>
    async function test() {
      await new Promise(loaded => window.addEventListener("load", loaded));
      const img = document.querySelector('img');
      const iframe = document.querySelector('iframe');
      const innerDoc = iframe.contentDocument || iframe.contentWindow.document;
      const div = innerDoc.querySelector('div');

      function onDropCallBack(event) {
        assert_equals(innerDoc.querySelector('div').textContent, 'PASS');
        return true;
      }
      dragDropTest(img, div, onDropCallBack,
        'Dragging the image to the vertical iframe scrollbar should copy it there', /*dragIframe=*/ null, iframe);
    }
    test();
  </script>
</body>
</html>
